improved merging algorithm

+++ python/pybirales/modules/detection/queue.py
@@ -10,6 +10,8 @@ class BeamCandidatesQueue:
         self._max_size = 20
         self.repository = None
 
+        self.candidates_to_delete = []
+
         # super(BeamCandidatesQueue, self).__init__(maxsize=self._max_size)
         # self.repository = BeamCandidateRepository()
 
@@ -18,22 +20,41 @@ class BeamCandidatesQueue:
 
     def enqueue(self, new_cluster):
         # Check if beam candidate is similar to candidate that was already added to the queue
+        # all_clusters_compared = False
+        # beam_queue = self.queue[new_cluster.beam_id]
+        # n = 0
+        # while True:
+        #     queue_length = len(self.queue[new_cluster.beam_id])
+        #
+        #     if n == queue_length:
+        #         break
+
         beam_queue = self.queue[new_cluster.beam_id]
-        for old_cluster in beam_queue:
+        log.warning('Queue %s length is %s (before merging)', new_cluster.beam_id, len(beam_queue))
+        for i, old_cluster in enumerate(beam_queue):
+            if old_cluster.to_delete:
+                continue
 
             if new_cluster.is_similar_to(old_cluster, threshold=0.2):
-                log.debug('Merging clusters in beam_queue %s (length: %s)', new_cluster.beam_id, len(beam_queue))
+                log.debug('Merging clusters [%s -> %s] in beam_queue %s (length: %s)',
+                          len(old_cluster.merge(new_cluster).time_data), len(old_cluster.time_data),
+                          new_cluster.beam_id, len(beam_queue))
                 # mark old cluster as 'to delete'
                 old_cluster.delete()
-
+                # beam_queue.remove(old_cluster)
                 # merge new cluster with the old one and add to the queue
-                beam_queue.insert(0, old_cluster.merge(new_cluster))
-
+                # beam_queue.insert(0, old_cluster.merge(new_cluster))
+                # log.warning('Inserted a merged cluster of size %s (from %s)',
+                #             len(old_cluster.merge(new_cluster).time_data), len(old_cluster.time_data))
+                self.enqueue(old_cluster.merge(new_cluster))
                 # Exit the loop once we added the new cluster to the queue
                 break
         else:
             # If we didn't merge, add the candidate to the queue
             beam_queue.insert(0, new_cluster)
+            log.warning('Inserted a cluster of size %s', len(new_cluster.time_data))
+
+        log.warning('Queue %s length is %s (after merging)', new_cluster.beam_id, len(beam_queue))
         # log.debug('Enqueuing took %1.3f s', time.time() - t1)
         # Pop the last element if queue is full
         self.dequeue(new_cluster.beam_id)
@@ -41,6 +62,9 @@ class BeamCandidatesQueue:
         if settings.detection.save_candidates:
             # Persist beam candidates
             self.save()
+
+        log.warning('Queue %s length is %s (after merging + deletion)', new_cluster.beam_id,
+                    len(self.queue[new_cluster.beam_id]))
 
     def dequeue(self, beam_id):
         """
@@ -70,3 +94,7 @@ class BeamCandidatesQueue:
         log.info('Added %s candidates, Deleted %s candidates',
                  len(candidates_to_save),
                  len(candidates_to_delete))
+
+        # Garbage collect - remove candidates marked for deletion from queue
+        for q, queue in enumerate(self.queue):
+            self.queue[q] = [candidate for candidate in queue if not candidate.to_delete]
